✨ refactor code
