+++ src/logic/useApi.ts
@@ -1,8 +1,9 @@
 import axios from 'axios'
 import { useStorage } from '@vueuse/core'
 
+const isDev = process.env.NODE_ENV === 'development'
 export const api = axios.create({
-  baseURL: 'http://localhost:3000',
+  baseURL: isDev ? 'http://localhost:3000' : ''
   timeout: 10000,
 })
